pptx_pres/requirements.txt
geopandas==0.12.2
matplotlib==3.5.1
pandas==1.4.2
plotly==5.6.0
python_pptx==0.6.21
